# cinema-booking-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cinema Booking API",
        "description": "Swagger documentation for Cinema Booking API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/movies/{id}": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Get a movie by ID",
                "operationId": "0665e4e5eeb4738291d9733b6015218b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "ulid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movie found"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
...
}